Refactor project structure and fix path issues
