+++ backend/main.py
@@ -1,0 +1,77 @@
+# main.py
+from fastapi import FastAPI, HTTPException
+from fastapi.middleware.cors import CORSMiddleware
+from pydantic import BaseModel, Field, validator
+import base64
+import os
+import requests
+import joblib
+import pandas as pd
+import numpy as np
+
+BASE_DIR = os.path.dirname(os.path.abspath(__file__))
+
+app = FastAPI()
+
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=["*"],
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"],
+)
+
+model = joblib.load(os.path.join(BASE_DIR, "yield_model.joblib"))
+
+########################
+# Crop Yield Prediction
+########################
+
+class PredictRequest(BaseModel):
+    Crop: str
+    CropCoveredArea: float = Field(..., gt=0)
+    CHeight: int = Field(..., ge=0)
+    CNext: str
+    CLast: str
+    CTransp: str
+    IrriType: str
+    IrriSource: str
+    IrriCount: int = Field(..., ge=1)
+    WaterCov: int = Field(..., ge=0, le=100)
+    Season: str
+
+class PredictResponse(BaseModel):
+    predicted_ExpYield: float
+
+@app.post("/predict", response_model=PredictResponse)
+async def predict_yield(data: PredictRequest):
+    try:
+        input_data = {
+            'Crop': data.Crop,
+            'CropCoveredArea': data.CropCoveredArea,
+            'CHeight': data.CHeight,
+            'CNext': data.CNext,
+            'CLast': data.CLast,
+            'CTransp': data.CTransp,
+            'IrriType': data.IrriType,
+            'IrriSource': data.IrriSource,
+            'IrriCount': data.IrriCount,
+            'WaterCov': data.WaterCov,
+            'Season': data.Season
+        }
+        df = pd.DataFrame([input_data])
+        
+        dummy_cols = ['Crop', 'CNext', 'CLast', 'CTransp', 'IrriType', 'IrriSource', 'Season']
+        df = pd.get_dummies(df, columns=dummy_cols, drop_first=True)
+        
+        feature_cols = list(model.get_booster().feature_names)
+        for col in feature_cols:
+            if col not in df.columns:
+                df[col] = 0
+        df = df[feature_cols]
+        
+        predicted_yield = model.predict(df)[0]
+        return {"predicted_ExpYield": float(predicted_yield)}
+    except Exception as e:
+        raise HTTPException(status_code=400, detail=str(e))
+
