--- frontend/How.css
@@ -1,138 +1,0 @@
-.howitworks {
-  max-width: 1200px;
-  margin: auto;
-  padding: 80px 20px;
-  text-align: center;
-  font-family: "Poppins", sans-serif;
-  background: linear-gradient(135deg, #e0f7fa, #e6fffa, #f0fdf4);
-  border-radius: 16px;
-  position: relative;
-  overflow: hidden;
-}
-
-/* Decorative floating glow */
-.howitworks::before {
-  content: "";
-  position: absolute;
-  top: -50px;
-  left: -50px;
-  width: 200px;
-  height: 200px;
-  background: radial-gradient(circle, #22c55e55, transparent 70%);
-  border-radius: 50%;
-  animation: float 6s ease-in-out infinite;
-}
-
-.howitworks-header h1 {
-  font-size: 3rem;
-  font-weight: 700;
-  margin-bottom: 12px;
-
-  /* fallback color */
-  color: #064e3b;
-
-  /* gradient text */
-  background-image: linear-gradient(90deg, #22c55e, #0ea5e9);
-  background-clip: text;
-  -webkit-background-clip: text; /* Safari/Chrome */
-  color: transparent;
-  -webkit-text-fill-color: transparent; /* Safari */
-}
-
-
-.howitworks-header p {
-  font-size: 1.2rem;
-  color: #333;
-  margin-bottom: 50px;
-}
-
-/* Steps Grid */
-.steps {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(260px, 1fr));
-  gap: 24px;
-}
-
-.step-card {
-  background: rgba(255, 255, 255, 0.85);
-  backdrop-filter: blur(12px);
-  border-radius: 16px;
-  padding: 30px 20px;
-  box-shadow: 0 6px 18px rgba(0, 0, 0, 0.08);
-  transition: transform 0.35s ease, box-shadow 0.35s ease;
-  position: relative;
-  overflow: hidden;
-  border: 1px solid rgba(34, 197, 94, 0.2);
-}
-
-.step-card:hover {
-  transform: translateY(-10px) scale(1.05);
-  box-shadow: 0 10px 25px rgba(0, 0, 0, 0.15);
-  border-color: #22c55e;
-}
-
-/* Step Number Badge */
-.step-card::before {
-  content: attr(data-step);
-  position: absolute;
-  top: -15px;
-  right: -15px;
-  background: #22c55e;
-  color: white;
-  font-weight: bold;
-  font-size: 1rem;
-  padding: 8px 12px;
-  border-radius: 50%;
-  box-shadow: 0 3px 8px rgba(0, 0, 0, 0.2);
-}
-
-/* Icons */
-.step-icon {
-  font-size: 2.5rem;
-  margin-bottom: 15px;
-  display: inline-block;
-  animation: bounce 2s infinite;
-}
-
-.step-card h3 {
-  font-size: 1.3rem;
-  font-weight: 600;
-  margin-bottom: 10px;
-  color: #064e3b;
-}
-
-.step-card p {
-  font-size: 1rem;
-  color: #444;
-  line-height: 1.5;
-}
-
-/* Animations */
-@keyframes bounce {
-  0%, 100% { transform: translateY(0); }
-  50% { transform: translateY(-8px); }
-}
-
-@keyframes float {
-  0%, 100% { transform: translate(0, 0); }
-  50% { transform: translate(20px, 20px); }
-}
-
-.fade-up {
-  animation: fadeUp 1s ease forwards;
-  opacity: 0;
-}
-
-@keyframes fadeUp {
-  from { transform: translateY(30px); opacity: 0; }
-  to { transform: translateY(0); opacity: 1; }
-}
-.slide-in {
-  animation: slideIn 1s ease forwards;
-  opacity: 0;
-}
-
-@keyframes slideIn {
-  from { transform: translateX(-100%); opacity: 0; }
-  to { transform: translateX(0); opacity: 1; }
-}
