+++ frontend/css/WeatherCard.css
@@ -1,0 +1,228 @@
+/* Overlay */
+.weather-overlay {
+  position: fixed;
+  top: 0; left: 0;
+  width: 100%; height: 100%;
+  background: rgba(0, 0, 0, 0.6);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 2000;
+  backdrop-filter: blur(6px);
+  animation: fadeIn 0.3s ease-in-out;
+}
+
+/* Popup container */
+.weather-popup {
+  position: relative;
+  background: rgba(255, 255, 255, 0.1);
+  backdrop-filter: blur(16px);
+  border-radius: 20px;
+  padding: 25px;
+  max-width: 380px;
+  width: 90%;
+  text-align: center;
+  color: #fff;
+  box-shadow: 0 10px 30px rgba(0,0,0,0.3);
+  animation: slideUp 0.4s ease-in-out;
+}
+
+/* Close button */
+.close-btn {
+  position: absolute;
+  top: 12px;
+  right: 12px;
+  background: rgba(0, 0, 0, 0.2);
+  border: none;
+  color: white;
+  font-size: 20px;
+  border-radius: 50%;
+  cursor: pointer;
+  padding: 6px 10px;
+  transition: 0.3s;
+}
+
+.close-btn:hover {
+  background: rgba(0, 0, 0, 0.4);
+  color: red;
+}
+
+/* Weather card content */
+.weather-card h2 {
+  margin-bottom: 5px;
+}
+
+.subtitle {
+  font-size: 14px;
+  opacity: 0.8;
+  margin-bottom: 15px;
+}
+
+.input-group {
+  display: flex;
+  gap: 8px;
+  margin-bottom: 12px;
+}
+
+.input-group input {
+  flex: 1;
+  padding: 10px;
+  border-radius: 10px;
+  border: none;
+  outline: none;
+}
+
+.input-group button,
+.location-btn {
+  padding: 10px 15px;
+  border-radius: 10px;
+  border: none;
+  cursor: pointer;
+  transition: 0.2s;
+}
+
+.input-group button {
+  background: #007bff;
+  color: #fff;
+}
+.input-group button:hover { background: #0056b3; }
+
+.location-btn {
+  background: #28a745;
+  color: #fff;
+}
+.location-btn:hover { background: #1e7e34; }
+
+/* Weather info */
+.weather-info {
+  margin-top: 20px;
+  background: rgba(255,255,255,0.15);
+  padding: 15px;
+  border-radius: 15px;
+}
+
+.icon-temp {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 12px;
+}
+
+.icon-temp h1 { font-size: 42px; }
+
+.desc {
+  margin: 8px 0;
+  text-transform: capitalize;
+  font-style: italic;
+}
+
+.details {
+  font-size: 14px;
+  margin-top: 8px;
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
+
+/* Animations */
+@keyframes fadeIn {
+  from { opacity: 0; transform: scale(0.9); }
+  to { opacity: 1; transform: scale(1); }
+}
+
+@keyframes slideUp {
+  from { transform: translateY(40px); opacity: 0; }
+  to { transform: translateY(0); opacity: 1; }
+}
+/* 📱 Mobile (up to 600px) */
+@media (max-width: 600px) {
+  .weather-popup {
+    max-width: 90%;
+    padding: 18px;
+    border-radius: 14px;
+  }
+
+  .weather-card h2 {
+    font-size: 1.4rem;
+  }
+
+  .subtitle {
+    font-size: 12px;
+  }
+
+  .input-group {
+    flex-direction: column;
+    gap: 10px;
+  }
+
+  .input-group input,
+  .input-group button,
+  .location-btn {
+    width: 100%;
+    font-size: 0.9rem;
+    padding: 10px;
+  }
+
+  .icon-temp h1 {
+    font-size: 32px;
+  }
+
+  .details {
+    font-size: 12px;
+  }
+}
+
+/* 📱 Small tablets (601px–900px) */
+@media (min-width: 601px) and (max-width: 900px) {
+  .weather-popup {
+    max-width: 420px;
+    padding: 22px;
+  }
+
+  .weather-card h2 {
+    font-size: 1.6rem;
+  }
+
+  .subtitle {
+    font-size: 13px;
+  }
+
+  .input-group input,
+  .input-group button,
+  .location-btn {
+    font-size: 1rem;
+    padding: 12px;
+  }
+
+  .icon-temp h1 {
+    font-size: 36px;
+  }
+
+  .details {
+    font-size: 13px;
+  }
+}
+
+/* 💻 Large screens (above 1200px) */
+@media (min-width: 1200px) {
+  .weather-popup {
+    max-width: 500px;
+    padding: 30px;
+  }
+
+  .weather-card h2 {
+    font-size: 1.8rem;
+  }
+
+  .subtitle {
+    font-size: 15px;
+  }
+
+  .icon-temp h1 {
+    font-size: 48px;
+  }
+
+  .details {
+    font-size: 15px;
+  }
+}
