+++ requirements.txt
@@ -13,4 +13,6 @@ joblib==1.3.2
 pydantic==2.6.2
 python-multipart==0.0.6
 
+requests
+xgboost
